Update 02-types.conf
Fixes issue with suricata (#236) where the value of _gateway is being applied to the host and in turn the observer.ip but the observer.ip field must be an IP address which results in an error (doesn't send the event to elasticsearch).  

Fix: Adjusted the host to rename to observer.hostname, for now.